fix: cart

+++ resources/views/livewire/features/cashier/cart.blade.php
@@ -1,108 +1,60 @@
-<section
-    id="section-cashier-cart"
-    data-component-id="cashier-cart"
-    class="flex-1"
->
-    <div class="w-full flex justify-end mb-3"
-         x-data="{
-            initIcons() {
-               setTimeout(() => { lucide.createIcons(); }, 0);
-            }
-        }"
-         x-init="initIcons()"
-         x-effect="initIcons()"
-    >
+<section id="section-cashier-cart" data-component-id="cashier-cart" class="flex-1">
+    <div class="w-full flex justify-end mb-3" x-data="{
+        initIcons() {
+            setTimeout(() => { lucide.createIcons(); }, 0);
+        }
+    }" x-init="initIcons()" x-effect="initIcons()">
         <div class="flex gap-1">
             <div class="relative group">
                 <div wire:ignore class="h-full flex items-center px-[0.5rem] absolute inset-y-0 start-0">
                     <i data-lucide="scan-barcode"
-                       class="text-neutral-500 group-focus-within:text-neutral-900 h-4 aspect-[1/1]"></i>
+                        class="text-neutral-500 group-focus-within:text-neutral-900 h-4 aspect-[1/1]"></i>
                 </div>
-                <input
-                    x-model="$store.cartStore.plu"
-                    placeholder="scan barcode"
+                <input x-model="$store.cartStore.plu" placeholder="scan barcode"
                     class="w-full text-xs ps-[2.05rem] pe-[0.825rem] py-[0.525rem] rounded text-neutral-700 border border-neutral-300 outline-none focus:outline-none focus:ring-0 focus:border-neutral-500 transition duration-300 ease-in"
-                    x-on:keydown.enter="$store.cartStore.findByPLU()"
-                />
+                    x-on:keydown.enter="$store.cartStore.findByPLU()" />
             </div>
-            <x-gxui.button.button
-                wire:ignore
-                x-on:click="$store.cartStore.searchProduct()"
-                class="!w-fit"
-                x-bind:disabled="false"
-            >
+            <x-gxui.button.button wire:ignore x-on:click="$store.cartStore.searchProduct()" class="!w-fit"
+                x-bind:disabled="false">
                 <div wire:ignore>
-                    <i data-lucide="box"
-                       class="text-white h-3 aspect-[1/1]"></i>
+                    <i data-lucide="box" class="text-white h-3 aspect-[1/1]"></i>
                 </div>
             </x-gxui.button.button>
         </div>
     </div>
-    <x-gxui.table.dynamic.table
-        store="cartStore"
-        dispatcher="getCart"
-        stateLoader="loading"
-    >
+    <x-gxui.table.dynamic.table store="cartStore" dispatcher="getCart" stateLoader="loading">
         <x-slot name="header">
-            <x-gxui.table.dynamic.th
-                class="flex-1 min-w-[150px]"
-            >
+            <x-gxui.table.dynamic.th class="flex-1 min-w-[150px]">
                 <span>Product</span>
             </x-gxui.table.dynamic.th>
-            <x-gxui.table.dynamic.th
-                contentClass="justify-center"
-                class="w-[80px]"
-            >
+            <x-gxui.table.dynamic.th contentClass="justify-center" class="w-[80px]">
                 <span>Unit</span>
             </x-gxui.table.dynamic.th>
-            <x-gxui.table.dynamic.th
-                contentClass="justify-end"
-                class="w-[100px]"
-            >
+            <x-gxui.table.dynamic.th contentClass="justify-end" class="w-[100px]">
                 <span>Price (Rp.)</span>
             </x-gxui.table.dynamic.th>
-            <x-gxui.table.dynamic.th
-                contentClass="justify-center"
-                class="w-[80px]"
-            >
+            <x-gxui.table.dynamic.th contentClass="justify-center" class="w-[80px]">
                 <span>Qty</span>
             </x-gxui.table.dynamic.th>
-            <x-gxui.table.dynamic.th
-                contentClass="justify-end"
-                class="w-[100px]"
-            >
+            <x-gxui.table.dynamic.th contentClass="justify-end" class="w-[100px]">
                 <span>Total (Rp)</span>
             </x-gxui.table.dynamic.th>
-            <x-gxui.table.dynamic.th
-                contentClass="justify-center"
-                class="w-[50px]"
-            >
+            <x-gxui.table.dynamic.th contentClass="justify-center" class="w-[50px]">
             </x-gxui.table.dynamic.th>
         </x-slot>
         <x-slot name="rows">
             <x-gxui.table.dynamic.row>
-                <x-gxui.table.dynamic.td
-                    class="flex-1 min-w-[150px]"
-                >
+                <x-gxui.table.dynamic.td class="flex-1 min-w-[150px]">
                     <span x-text="data.name"></span>
                 </x-gxui.table.dynamic.td>
-                <x-gxui.table.dynamic.td
-                    contentClass="justify-center"
-                    class="w-[80px]"
-                >
+                <x-gxui.table.dynamic.td contentClass="justify-center" class="w-[80px]">
                     <span x-text="data.unit"></span>
                 </x-gxui.table.dynamic.td>
-                <x-gxui.table.dynamic.td
-                    contentClass="justify-end"
-                    class="w-[100px]"
-                >
+                <x-gxui.table.dynamic.td contentClass="justify-end" class="w-[100px]">
                     <span x-text="data.price.toLocaleString('id-ID') ?? 0"></span>
                 </x-gxui.table.dynamic.td>
-                <x-gxui.table.dynamic.td
-                    contentClass="justify-center"
-                    class="w-[80px]"
-                >
-                    <div x-data="{error: false}">
+                <x-gxui.table.dynamic.td contentClass="justify-center" class="w-[80px]">
+                    <div x-data="{ error: false }">
                         <input
                             class="w-10 px-1 py-1 text-xs text-center rounded text-neutral-700 border border-neutral-300 outline-none focus:outline-none focus:ring-0 focus:border-neutral-500 transition duration-300 ease-in"
                             x-model="data.qty"
@@ -110,45 +62,30 @@
                                 $store.cartStore.updateCart(data.id, $event.target.value);
                                 error = $event.target.value.trim() === '';
                             "
-                            :class="error ? 'border-red-500 focus:border-red-500' : 'border-neutral-300 focus:border-neutral-500'"
-                        />
+                            :class="error ? 'border-red-500 focus:border-red-500' :
+                                'border-neutral-300 focus:border-neutral-500'" />
                     </div>
                 </x-gxui.table.dynamic.td>
-                <x-gxui.table.dynamic.td
-                    contentClass="justify-end"
-                    class="w-[100px]"
-                >
+                <x-gxui.table.dynamic.td contentClass="justify-end" class="w-[100px]">
                     <span x-text="data.total.toLocaleString('id-ID') ?? 0"></span>
                 </x-gxui.table.dynamic.td>
-                <x-gxui.table.dynamic.td
-                    contentClass="justify-center"
-                    class="w-[50px]"
-                    x-data="{
-                        initIcons() {
-                           setTimeout(() => { lucide.createIcons(); }, 0);
-                        }
-                    }"
-                    x-init="initIcons()"
-                    x-effect="initIcons()"
-                >
-                    <div
-                        class="cursor-pointer w-fit"
-                        wire:ignore
-                        x-on:click="$store.cartStore.confirmDelete(index)"
-                    >
+                <x-gxui.table.dynamic.td contentClass="justify-center" class="w-[50px]" x-data="{
+                    initIcons() {
+                        setTimeout(() => { lucide.createIcons(); }, 0);
+                    }
+                }"
+                    x-init="initIcons()" x-effect="initIcons()">
+                    <div class="cursor-pointer w-fit" wire:ignore x-on:click="$store.cartStore.confirmDelete(index)">
                         <i data-lucide="trash"
-                           class="text-red-500 group-focus-within:text-neutral-900 h-3 aspect-[1/1]">
+                            class="text-red-500 group-focus-within:text-neutral-900 h-3 aspect-[1/1]">
                         </i>
                     </div>
                 </x-gxui.table.dynamic.td>
             </x-gxui.table.dynamic.row>
         </x-slot>
     </x-gxui.table.dynamic.table>
-    <x-gxui.modal.confirmation
-        store="cartStore"
-        dispatcher="$store.cartStore.deleteItem()"
-        onCloseDispatcher="$store.cartStore.closeDelete()"
-    ></x-gxui.modal.confirmation>
+    <x-gxui.modal.confirmation store="cartStore" dispatcher="$store.cartStore.deleteItem()"
+        onCloseDispatcher="$store.cartStore.closeDelete()"></x-gxui.modal.confirmation>
 </section>
 
 @push('scripts')
@@ -156,7 +93,7 @@
         document.addEventListener('alpine:init', () => {
             Alpine.store('cartStore', {
                 component: null,
-                transactionStore: null,
+                actionLoaderStore: null,
                 billingStore: null,
                 toastStore: null,
                 searchStore: null,
@@ -165,16 +102,20 @@
                 loading: true,
                 data: [],
                 deleteIndex: 0,
-                init: function () {
-                    Livewire.hook('component.init', ({component}) => {
-                        const componentID = document.querySelector('[data-component-id="cashier-cart"]')?.getAttribute('wire:id');
+                init: function() {
+                    Livewire.hook('component.init', ({
+                        component
+                    }) => {
+                        const componentID = document.querySelector(
+                            '[data-component-id="cashier-cart"]')?.getAttribute('wire:id');
                         if (component.id === componentID) {
                             this.component = component;
                             this.billingStore = Alpine.store('billingStore');
-                            this.transactionStore = Alpine.store('transactionStore');
+                            this.actionLoaderStore = Alpine.store('gxuiActionLoader');
                             this.searchStore = Alpine.store('cartSearchStore');
                             this.toastStore = Alpine.store('gxuiToastStore');
-                            this.modalConfirmationStore = Alpine.store('gxuiModalConfirmationStore');
+                            this.modalConfirmationStore = Alpine.store(
+                                'gxuiModalConfirmationStore');
                         }
                     })
                 },
@@ -186,18 +127,22 @@
                         this._setTotal();
                         this._updateStorageCart();
                     } else {
-                        this.transactionStore.showLoading('find product...');
+                        this.actionLoaderStore.start('finding item..')
                         this.component.$wire.call('getProductByPLU', this.plu)
                             .then(response => {
-                                const {success, message, data} = response;
+                                const {
+                                    success,
+                                    message,
+                                    data
+                                } = response;
                                 if (success) {
                                     this.addToCart(data);
                                 } else {
                                     this.toastStore.failed(message);
                                 }
                             }).finally(() => {
-                            this.transactionStore.closeLoading();
-                        })
+                                this.actionLoaderStore.end();
+                            })
                     }
                     this.plu = '';
                 },
@@ -207,10 +152,13 @@
                 print() {
                     this.component.$wire.call('print')
                         .then(response => {
-                            const {success, message, data} = response;
+                            const {
+                                success,
+                                message,
+                                data
+                            } = response;
                             console.log(response)
-                        }).finally(() => {
-                    })
+                        }).finally(() => {})
                 },
                 addToCart(item) {
                     const plu = item['price_list_unit'];
